# api-documentation.yaml
openapi: 3.0.0
info:
  title: Party Picker API
  version: 1.0.0
...
security:
  - BearerAuth: []
  - BasicAuth: []
paths:
...
  /events:
    get:
      summary: Get events
      tags:
        - Events
      parameters:
        - in: query
          name: user
          schema:
            type: integer
          description: The ID of the user to retrieve events for
          required: true
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: The page number to retrieve
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: The number of events per page
        - in: query
          name: from
          schema:
            type: string
            format: date-time
          description: The minimum date of the events to retrieve
        - in: query
          name: to
          schema:
            type: string
            format: date-time
          description: The maximum date of the events to retrieve
      responses:
        "200":
          description: The events were successfully retrieved
          headers:
            first:
              $ref: "#/components/headers/first"
            last:
              $ref: "#/components/headers/last"
            next:
              $ref: "#/components/headers/next"
            prev:
              $ref: "#/components/headers/prev"
          content:
            application/vnd.events.v1+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "400":
          description: Invalid request
    post:
      summary: Create an event
      tags:
        - Events
      requestBody:
        description: The details of the new event
        required: true
        content:
          application/vnd.event.v1+json:
            schema:
              $ref: "#/components/schemas/Event"
      responses:
        "201":
          description: The event was successfully created
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
  /events/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: The ID of the event
    get:
      summary: Get an event by ID
      tags:
        - Events
      responses:
        "200":
          description: The event was successfully retrieved
          content:
            application/vnd.event.v1+json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: The event was not found
    put:
      summary: Update an event
      tags:
        - Events
      requestBody:
        description: The new event details
        required: true
        content:
          application/vnd.event.v1+json:
            schema:
              $ref: "#/components/schemas/Event"
      responses:
        "200":
          description: The event was successfully updated
          content:
            application/vnd.event+json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The event was not found
    delete:
      summary: Delete an event
      tags:
        - Events
      responses:
        "204":
          description: The event was successfully deleted
        "404":
          description: The event was not found
  /events/{id}/guests:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: The ID of the event
    get:
      summary: Get guests of an event
      tags:
        - Events
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: The page number to retrieve
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: The number of guests per page
      responses:
        "200":
          description: The guests were successfully retrieved
          headers:
            first:
              $ref: "#/components/headers/first"
            last:
              $ref: "#/components/headers/last"
            next:
              $ref: "#/components/headers/next"
            prev:
              $ref: "#/components/headers/prev"
          content:
            application/vnd.guests.v1+json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          description: The event was not found
    post:
      summary: Add a guest to an event
      tags:
        - Events
      requestBody:
        description: The details of the new guest
        required: true
        content:
          application/vnd.guest.v1+json:
            schema:
              $ref: "#/components/schemas/Guest"
      responses:
        "201":
          description: The guest was successfully added
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The event was not found
    patch:
      summary: Update a guest of an event
      tags:
        - Events
      requestBody:
        description: The new guest details
        required: true
        content:
          application/vnd.guestStatus.v1+json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/GuestStatusEnum"
      responses:
        "200":
          description: The guest was successfully updated
          content:
            application/vnd.guest+json:
              schema:
                $ref: "#/components/schemas/Guest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The event was not found
    delete:
      summary: Remove a guest from an event
      tags:
        - Events
      responses:
        "204":
          description: The guest was successfully removed
        "400":
          description: Invalid request
        "404":
          description: The event or guest was not found
  /images:
    post:
      summary: Upload an image
      requestBody:
        description: The image to upload
        required: true
        content:
          multipart/form-data:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: The image was successfully uploaded
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
  /images/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uri
        description: The URI of the image
    get:
      summary: Get an image by URI
      responses:
        "200":
          description: The image was successfully retrieved
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: The image was not found
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: The ID of the event
        name:
          type: string
          minLength: 4
          maxLength: 50
          description: The name of the event
        description:
          type: string
          description: The description of the event
        date:
          type: string
          format: date-time
          description: The date of the event
        numberOfGuests:
          type: integer
          minimum: 1
          description: The number of guests expected at the event
        district:
          type: string
          description: The location of the event
        self:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the full details of the event
        owner:
          type: string
          format: uri
          description: The URI of the organizer
        relations:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the relations of the event
        relatedOfferings:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the offerings related to the event
        reviews:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the reviews of the event
        guests:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the guests of the event
...
    Guest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: The email of the guest
        status:
          $ref: "#/components/schemas/GuestStatusEnum"
        self:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the full details of the guest
    GuestStatusEnum:
      type: string
      enum:
        - NEW
        - PENDING
        - ACCEPTED
        - REJECTED
      description: The status of the guest
    FormErrors:
      type: array
      items:
        type: object
        properties:
          fieldName:
            type: string
            description: The name of the field that caused the error
          message:
            type: string
            description: The error message
...
  headers:
    first:
      schema:
        type: string
        format: uri
      description: The URI to the first page of the results
    last:
      schema:
        type: string
        format: uri
      description: The URI to the last page of the results
    next:
      schema:
        type: string
        format: uri
      description: The URI to the next page of the results, if there is one
    prev:
      schema:
        type: string
        format: uri
      description: The URI to the previous page of the results, if there is one
servers:
  - description: Party Picker API
    url: http://pawserver.it.itba.edu.ar/paw-2023b-04/api